requirements.txt
Flask==3.0.2
pytest==8.0.2
selenium==4.18.1
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
pytest-html==4.1.1
pytest-selenium==4.1.0
allure-pytest==2.13.2
gunicorn==21.2.0
webdriver-manager==4.0.1
coverage==7.4.3 